Rename inner TDI namespace from 'tdix' to 'tdi'
