+++ stratum/hal/lib/tdi/dpdk/dpdk_chassis_manager.h
@@ -22,7 +22,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 // Lock which protects chassis state across the entire switch.
 extern absl::Mutex chassis_lock;
@@ -227,7 +227,7 @@ class DpdkChassisManager {
   friend class DpdkChassisManagerTest;
 };
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
