+++ stratum/hal/lib/tdi/dpdk/dpdk_switch_test.cc
@@ -25,7 +25,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 using ::testing::_;
 using ::testing::DoAll;
@@ -652,6 +652,6 @@ TEST_F(DpdkSwitchTest, SetNoContentsPass) {
 // TODO(unknown): Complete unit test coverage.
 
 }  // namespace
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
