+++ stratum/hal/lib/tdi/tdi_chassis_manager.cc
@@ -27,7 +27,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 using PortStatusEvent = TdiSdeInterface::PortStatusEvent;
 using TransceiverEvent = PhalInterface::TransceiverEvent;
@@ -1373,6 +1373,6 @@ void TdiChassisManager::CleanupInternalState() {
   return status;
 }
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
