+++ stratum/hal/lib/tdi/tdi_chassis_manager_mock.h
@@ -14,7 +14,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 class TdiChassisManagerMock : public TdiChassisManager {
  public:
@@ -43,7 +43,7 @@ class TdiChassisManagerMock : public TdiChassisManager {
   MOCK_CONST_METHOD1(GetUnitFromNodeId, ::util::StatusOr<int>(uint64 node_id));
 };
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
