+++ stratum/hal/lib/tdi/tdi_chassis_manager_test.cc
@@ -29,7 +29,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 using PortStatusEvent = TdiSdeInterface::PortStatusEvent;
 using test_utils::EqualsProto;
@@ -838,6 +838,6 @@ TEST_F(TdiChassisManagerTest, VerifyChassisConfigSuccess) {
   ASSERT_OK(VerifyChassisConfig(config1));
 }
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
