+++ stratum/hal/lib/tdi/tdi_counter_manager.cc
@@ -15,7 +15,7 @@ DECLARE_uint32(tdi_table_sync_timeout_ms);
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 std::unique_ptr<TdiCounterManager> TdiCounterManager::CreateInstance(
     TdiSdeInterface* tdi_sde_interface, int device) {
@@ -120,6 +120,6 @@ TdiCounterManager::TdiCounterManager(TdiSdeInterface* tdi_sde_interface,
   return ::util::OkStatus();
 }
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
