+++ stratum/hal/lib/tdi/tdi_counter_manager.h
@@ -20,7 +20,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 class TdiCounterManager {
  public:
@@ -60,7 +60,7 @@ class TdiCounterManager {
   const int device_;
 };
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
