+++ stratum/hal/lib/tdi/tdi_packetio_manager.cc
@@ -23,7 +23,7 @@ DECLARE_bool(incompatible_enable_tdi_legacy_bytestring_responses);
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 TdiPacketioManager::TdiPacketioManager(TdiSdeInterface* tdi_sde_interface,
                                        int device)
@@ -377,6 +377,6 @@ void* TdiPacketioManager::SdeRxThreadFunc(void* arg) {
   return nullptr;
 }
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
