+++ stratum/hal/lib/tdi/tdi_packetio_manager.h
@@ -24,7 +24,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 class TdiPacketioManager {
  public:
@@ -142,7 +142,7 @@ class TdiPacketioManager {
   friend class TdiPacketioManagerTest;
 };
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
