+++ stratum/hal/lib/tdi/tdi_pre_manager_test.cc
@@ -17,7 +17,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 using ::testing::_;
 using ::testing::DoAll;
@@ -81,6 +81,6 @@ TEST_F(TdiPreManagerTest, DeleteMulticastGroupTest) {
                                              ::p4::v1::Update::DELETE, entry));
 }
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
