+++ stratum/hal/lib/tdi/tdi_switch.h
@@ -18,7 +18,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 class TdiChassisManager;
 
@@ -131,7 +131,7 @@ class TdiSwitch : public SwitchInterface {
   std::map<uint64, TdiNode*> node_id_to_tdi_node_;  //  pointers not owned
 };
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
