+++ stratum/hal/lib/tdi/utils.h
@@ -19,7 +19,7 @@
 
 namespace stratum {
 namespace hal {
-namespace tdix {
+namespace tdi {
 
 // A set of helper functions to determine whether a P4 match object constructed
 // from a bfrt table key is a "don't care" match.
@@ -49,7 +49,7 @@ std::string RangeDefaultHigh(size_t bitwidth);
 // byte string.
 int NumBitsToNumBytes(int num_bits);
 
-}  // namespace tdix
+}  // namespace tdi
 }  // namespace hal
 }  // namespace stratum
